Create README.md
# say-chesse
PHP JS HTML5 capture gps ip and photos
Just copy and past the code on a web server. 
Change the index.php header, the time in seconds for foward the page, and the place where it goes.
The web-site will take a photo, the gps location and the ip of the target.
You can perform better phishing sites putting these code in others phishing pages.
Code for study.I don't take responsability for ilegal porposes.